services/extract_validate/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
minio==7.2.9
pandas==2.2.3
openpyxl==3.1.5
python-dateutil==2.9.0.post0
rapidfuzz==3.9.6      # for name match score
pdfminer.six==20240706
python-docx==1.1.2
requests==2.32.3
grpcio==1.66.2
grpcio-tools==1.66.2